Crate pci

Structs

PciBus
A PCI bus, which contains a list of PCI devices on that bus.

PciDevice
Contains information common to every type of PCI Device, and offers functions for reading/writing to the PCI configuration space.

PciLocation
The bus, slot, and function number of a given PCI device. This offers methods for reading and writing the PCI config space.

Enums

PciConfigSpaceAccessMechanism
Lists the 2 possible PCI configuration space access mechanisms that can be found from the LSB of the devices’s BAR0

Constants

MSIX_CAPABILITY

MSI_CAPABILITY

PCI_BAR0

PCI_BAR1

PCI_BAR2

PCI_BAR3

PCI_BAR4

PCI_BAR5

PCI_BIST

PCI_CACHE_LINE_SIZE

PCI_CAPABILITIES

PCI_CARDBUS_CIS

PCI_CLASS

PCI_COMMAND

PCI_DEVICE_ID

PCI_EXPANSION_ROM_BASE

PCI_HEADER_TYPE

PCI_INTERRUPT_LINE

PCI_INTERRUPT_PIN

PCI_LATENCY_TIMER

PCI_MAX_LATENCY

PCI_MIN_GRANT

PCI_PROG_IF

PCI_REVISION_ID

PCI_STATUS

PCI_SUBCLASS

PCI_SUBSYSTEM_ID

PCI_SUBSYSTEM_VENDOR_ID

PCI_VENDOR_ID

Functions

get_pci_buses
Returns a list of all PCI buses in this system. If the PCI bus hasn’t been initialized, this initializes the PCI bus & scans it to enumerates devices.

get_pci_device_bsf
Returns a reference to the PciDevice with the given bus, slot, func identifier. If the PCI bus hasn’t been initialized, this initializes the PCI bus & scans it to enumerates devices.

pci_device_iter
Returns an iterator that iterates over all PciDevices, in no particular guaranteed order. If the PCI bus hasn’t been initialized, this initializes the PCI bus & scans it to enumerates devices.