Crate pci

Structs

MsixVectorEntry

A single Message Signaled Interrupt entry.

MsixVectorTable

A memory-mapped array of MsixVectorEntry

PciBus

A PCI bus, which contains a list of PCI devices on that bus.

PciDevice

Contains information common to every type of PCI Device, and offers functions for reading/writing to the PCI configuration space.

PciLocation

The bus, slot, and function number of a given PCI device. This offers methods for reading and writing the PCI config space.

Enums

InterruptPin

PciCapability

PciConfigSpaceAccessMechanism

Lists the 2 possible PCI configuration space access mechanisms that can be found from the LSB of the devices’s BAR0

Functions

get_pci_buses

Returns a list of all PCI buses in this system. If the PCI bus hasn’t been initialized, this initializes the PCI bus & scans it to enumerates devices.

get_pci_device_bsf

Returns a reference to the PciDevice with the given bus, slot, func identifier. If the PCI bus hasn’t been initialized, this initializes the PCI bus & scans it to enumerates devices.

pci_device_iter

Returns an iterator that iterates over all PciDevices, in no particular guaranteed order. If the PCI bus hasn’t been initialized, this initializes the PCI bus & scans it to enumerates devices.