Struct pci::PciDevice

pub struct PciDevice {

    pub location: PciLocation,
    pub class: u8,
    pub subclass: u8,
    pub prog_if: u8,
    pub bars: [u32; 6],
    pub vendor_id: u16,
    pub device_id: u16,
    pub command: u16,
    pub status: u16,
    pub revision_id: u8,
    pub cache_line_size: u8,
    pub latency_timer: u8,
    pub header_type: u8,
    pub bist: u8,
    pub int_pin: u8,
    pub int_line: u8,

}

Contains information common to every type of PCI Device, and offers functions for reading/writing to the PCI configuration space.

For more, see this partial table of class, subclass, and prog_if codes,

Fields
location: PciLocation

the bus, slot, and function number that locates this PCI device in the bus tree.

class: u8

The class code, used to determine device type.

subclass: u8

The subclass code, used to determine device type.

prog_if: u8

The programming interface of this PCI device, also used to determine device type.

bars: [u32; 6]

The six Base Address Registers (BARs)

vendor_id: u16device_id: u16command: u16status: u16revision_id: u8cache_line_size: u8latency_timer: u8header_type: u8bist: u8int_pin: u8int_line: u8
Implementations

impl PciDevice

pub fn determine_mem_base(
&self,
bar_index: usize
) -> Result<PhysicalAddress, &'static str>

Returns the base address of the memory region specified by the given BAR (Base Address Register) for this PCI device.

Argument

bar_index must be between 0 and 5 inclusively, as each PCI device can only have 6 BARs at the most.

Note that if the given BAR actually indicates it is part of a 64-bit address, it will be used together with the BAR right above it (bar + 1), e.g., BAR1:BAR0. If it is a 32-bit address, then only the given BAR will be accessed.

TODO: currently we assume the BAR represents a memory space (memory mapped I/O) rather than I/O space like Port I/O. Obviously, this is not always the case. Instead, we should return an enum specifying which kind of memory space the calculated base address is.

pub fn determine_mem_size(&self, bar_index: usize) -> u32

Returns the size in bytes of the memory region specified by the given BAR (Base Address Register) for this PCI device.

Argument

bar_index must be between 0 and 5 inclusively, as each PCI device can only have 6 BARs at the most.

pub fn pci_enable_msi(
&self,
core_id: u8,
int_num: u8
) -> Result<(), &'static str>

Enable MSI interrupts for a PCI device. We assume the device only supports one MSI vector and set the interrupt number and core id for that vector. If the MSI capability is not supported then an error message is returned.

Arguments

core_id: core that interrupt will be routed to

int_num: interrupt number to assign to the MSI vector

pub fn pci_enable_msix(&self) -> Result<(), &'static str>

Enable MSI-X interrupts for a PCI device. Only the enable bit is set and the remaining initialization steps of setting the interrupt number and core id should be completed in the device driver.