Struct pci::PciLocation

pub struct PciLocation { /* fields omitted */ }

The bus, slot, and function number of a given PCI device. This offers methods for reading and writing the PCI config space.

Implementations

impl PciLocation

pub fn bus(&self) -> u16

pub fn slot(&self) -> u16

pub fn function(&self) -> u16

pub fn pci_read_32(&self, offset: u16) -> u32

read 32-bit data at the specified offset from the PCI device specified by the given bus, slot, func set.

pub fn pci_read_16(&self, offset: u16) -> u16

Read 16-bit data at the specified offset from this PCI device.

pub fn pci_read_8(&self, offset: u16) -> u8

Read 8-bit data at the specified offset from the PCI device.

pub fn pci_write(&self, offset: u16, value: u32)

Write 32-bit data to the specified offset for the PCI device.

pub fn pci_set_command_bus_master_bit(&self)

Sets the PCI device’s bit 3 in the command portion, which is apparently needed to activate DMA (??)

pub fn pci_set_interrupt_disable_bit(&self)

Sets the PCI device’s command bit 10 to disable legacy interrupts

pub fn find_pci_capability(&self, pci_capability: u16) -> Option<u16>

Explores the PCI config space and returns address of requested capability, if present. PCI capabilities are stored as a linked list in the PCI config space, with each capability storing the pointer to the next capability right after its ID. The function returns a None value if capabilities are not valid for this device or if the requested capability is not present.